Fix prototype typo.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@3485 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/math/fmod.c
@@ -3,7 +3,7 @@
  * This file is part of the w64 mingw-runtime package.
  * No warranty is given; refer to the file DISCLAIMER.PD within this package.
  */
-double fmodl (double x, double y);
+double fmod (double x, double y);
 
 double
 fmod (double x, double y)
